--- sip-trunk-provisioning-tool/nodejs/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "trunk-creation-tutorial",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Russelle",
-  "license": "MIT",
-  "dependencies": {
-    "express": "^4.17.1",
-    "purecloud-platform-client-v2": "^77.0.0"
-  },
-  "devDependencies": {
-    "eslint": "^6.8.0"
-  }
-}
